Fix compile time issues with `Router::boxed`
